LayoutTests/fast/canvas/webgl/context-lost-restored-p3-expected.txt

Restored context should keep colorspace.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test losing and restoring a context.
Test valid context
PASS gl.isContextLost() is false
PASS gl.getError() is gl.NO_ERROR
PASS gl.drawingBufferColorSpace is "display-p3"

PASS gl.isContextLost() is true
PASS gl.getError() is gl.CONTEXT_LOST_WEBGL
PASS gl.getError() is gl.NO_ERROR
PASS contextLostEventFired is false
PASS gl.drawingBufferColorSpace is "display-p3"
Test lost context
PASS contextLostEventFired is false
PASS gl.isContextLost() is true
PASS gl.getError() is gl.NO_ERROR

PASS extension.restoreContext() was expected value: NO_ERROR.
PASS gl.isContextLost() is true
PASS gl.getError() is gl.NO_ERROR
Test restored context
PASS contextRestoredEventFired is false
PASS gl.isContextLost() is false
PASS gl.getError() is gl.NO_ERROR
PASS gl.drawingBufferColorSpace is "display-p3"

PASS successfullyParsed is true

TEST COMPLETE
